Menu lidar listo

+++ Screen/screen platformio/include/display.hpp
@@ -18,7 +18,7 @@ namespace display {
     */
 
     void escribirTexto(LCDWIKI_KBV mylcd, uint16_t color, uint8_t tamanno, String texto,
-    uint8_t coordenada_X, uint8_t coordenada_Y);
+    int coordenada_X, int coordenada_Y);
 
     //Si no hay interrupciones devuelve 0, si las hay devuelve el codigo del menu que hay que ejecutar
     int escribirErrores(LCDWIKI_KBV mylcd, uint16_t color, uint8_t tamanno, String texto,
